Rename middleware from camel case to snake case.
